rename : 공통 속성 entity common 패키지로 이동(#4)
